Added double conversion.

+++ doom_nukem.h
@@ -86,7 +86,7 @@ void			draw_minimap(t_xy size, t_xy pos, t_xy dir, t_home *home);
 void			draw_grid(t_frame *frame);
 int				update_sector(t_home *home);
 void			translate_world_view(t_home *home, t_xy step);
-void			transform_world_view(t_home *home, float delta_dir);
+void			transform_world_view(t_home *home, double delta_dir);
 void			perspective_transformation(t_point *x0, t_point *x1, t_home *home, int idx);
 
 void			draw_text(t_home *home, char *text, t_frame *frame, t_xy pos);
