+++ srcs/vec_math/vec2_a.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   vec2_a.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmaarela <tmaarela@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/05 19:15:12 by tmaarela          #+#    #+#             */
-/*   Updated: 2020/11/05 19:16:59 by tmaarela         ###   ########.fr       */
+/*   Updated: 2021/02/19 13:16:12 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
-t_xy			vec2(float x, float y)
+t_xy			vec2(double x, double y)
 {
 	return ((t_xy){x, y});
 }
