+++ srcs/vec_math/vec2_b.c
@@ -6,13 +6,13 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/05 19:16:10 by tmaarela          #+#    #+#             */
-/*   Updated: 2021/02/02 16:37:09 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 13:16:38 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../doom_nukem.h"
 
-t_xy			vec2_rot(t_xy v, float angle)
+t_xy			vec2_rot(t_xy v, double angle)
 {
 	return ((t_xy){
 		v.x * cos(angle) - v.y * sin(angle),
@@ -22,7 +22,7 @@ t_xy			vec2_rot(t_xy v, float angle)
 
 t_xy			vec2_norm(t_xy v)
 {
-	float mag;
+	double mag;
 
 	mag = sqrt((v.x * v.x) + (v.y * v.y));
 	return ((t_xy){
@@ -31,7 +31,7 @@ t_xy			vec2_norm(t_xy v)
 	});
 }
 
-t_xy			vec2_mul(t_xy v, float scalar)
+t_xy			vec2_mul(t_xy v, double scalar)
 {
 	return ((t_xy){
 		v.x * scalar,
@@ -39,22 +39,22 @@ t_xy			vec2_mul(t_xy v, float scalar)
 	});
 }
 
-float			vec2_dot(t_xy a, t_xy b)
+double			vec2_dot(t_xy a, t_xy b)
 {
 	return (a.x * b.x + a.y * b.y);
 }
 
-float			vec2_mag(t_xy a)
+double			vec2_mag(t_xy a)
 {
 	return sqrt(a.x * a.x + a.y * a.y);
 }
 
-float			vec2_angle(t_xy a, t_xy b)
+double			vec2_angle(t_xy a, t_xy b)
 {
 	return acos(vec2_dot(a, b) / (vec2_mag(a) * vec2_mag(b)));
 }
 
-float			ft_map(float in, t_range from, t_range to)
+double			ft_map(double in, t_range from, t_range to)
 {
 	return (to.x + ((in - from.x) * (to.y - to.x)) / (from.y - from.x));
 }
